app/requirements.txt
alembic
Babel
click
Flask
Flask-Migrate
Flask-Moment
Flask-SQLAlchemy
Flask-WTF
itsdangerous
Jinja2
Mako
MarkupSafe
python-dateutil
python-editor
pytz
six
SQLAlchemy
Werkzeug
WTForms